lab2 fixed

+++ src/lab2/Calculator.java
@@ -19,11 +19,11 @@ public class Calculator {
     public static double multiply(double arg1, double arg2) {
         return arg1 * arg2 ;
     }
-    public static int divide(int arg1, int arg2) {
-        return arg1 / arg2 ;
+    public static double divide(int arg1, int arg2) {
+        return (double) arg1 / arg2 ;
     }
-    public static long divide(long arg1, long arg2) {
-        return arg1 / arg2 ;
+    public static double divide(long arg1, long arg2) {
+        return (double) arg1 / arg2 ;
     }
     public static double divide(double arg1, double arg2) {
         return arg1 / arg2 ;
@@ -37,4 +37,5 @@ public class Calculator {
     public static double substract(double arg1, double arg2) {
         return arg1 - arg2 ;
     }
+
 }
